AFIRGEN FINAL/gguf model server/requirements.txt
fastapi==0.116.1
pydantic==2.11.7
uvicorn==0.35.0
llama-cpp-python==0.3.1  
filelock==3.16.1  